[multibody] Adjust contact results targets for uniformity
The contact results and its child structs are all one logical entity, so
they should be under a single build label.

Also fix a related unit test to be properly split up, per its TODO.

Deprecate the old bazel labels.
